+++ docs/release-notes/release-notes.md
@@ -6,6 +6,7 @@
 
 ## AtlasProject Resource
   * [3rd Party Integration](https://docs.atlas.mongodb.com/reference/api/third-party-integration-settings/) are supported `spec.integrations`
+  * [GCP Private Endpoints](https://www.mongodb.com/docs/atlas/reference/api/private-endpoints/) are now supported
 
 ## AtlasCluster Resource
 * Changes